+++ src/__generated__/QueryRenderersFairQuery.graphql.ts
@@ -30,6 +30,7 @@ query QueryRenderersFairQuery(
 fragment Fair_fair on Fair {
   ...FairHeader_fair
   id
+  hours
   location {
     ...LocationMap_location
     __id
@@ -137,7 +138,7 @@ return {
   "kind": "Request",
   "operationKind": "query",
   "name": "QueryRenderersFairQuery",
-  "id": "06148b7e73dd412041f7cce4d1a3ffbd",
+  "id": "2930829cfe475e9fd45d0cbc390cde2b",
   "text": null,
   "metadata": {},
   "fragment": {
@@ -281,6 +282,13 @@ return {
             "storageKey": null
           },
           v2,
+          {
+            "kind": "ScalarField",
+            "alias": null,
+            "name": "hours",
+            "args": null,
+            "storageKey": null
+          },
           {
             "kind": "LinkedField",
             "alias": null,